pyproject.toml
[project]
name = "hoopla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "nltk==3.9.1",
    "sentence-transformers",
    "numpy",
    "python-dotenv>=1.2.1",
    "google-genai>=1.53.0",
    "pillow>=12.0.0",
]
